Updates based on dev for Ludum Dare 42.

+++ VFrame/VText.cpp
@@ -344,7 +344,7 @@ void VText::SetWrap(const VTextWrap& wrapMode)
 	}
 }
 
-void VText::SetLineSpaceModifier(float lineSpacing)
+void VText::SetLineSpaceModifier(int lineSpacing)
 {
 	if (lineSpaceModifier != lineSpacing)
 	{
